+++ src/gameobjects/tilesprite/TileSprite.js
@@ -7,6 +7,7 @@
 var CanvasPool = require('../../display/canvas/CanvasPool');
 var Class = require('../../utils/Class');
 var Components = require('../components');
+var CONST = require('../../const');
 var GameObject = require('../GameObject');
 var GetPowerOfTwo = require('../../math/pow2/GetPowerOfTwo');
 var TileSpriteRender = require('./TileSpriteRender');
@@ -179,13 +180,17 @@ var TileSprite = new Class({
 
         this.updateTileTexture();
 
-        scene.sys.game.renderer.onContextRestored(function (renderer)
+        if (scene.sys.game.config.renderType === CONST.WEBGL)
         {
-            var gl = renderer.gl;
-            this.tileTexture = null;
-            this.dirty = true;
-            this.tileTexture = renderer.createTexture2D(0, gl.LINEAR, gl.LINEAR, gl.REPEAT, gl.REPEAT, gl.RGBA, this.canvasBuffer, this.potWidth, this.potHeight);
-        }, this);
+            scene.sys.game.renderer.onContextRestored(function (renderer)
+            {
+                var gl = renderer.gl;
+
+                this.tileTexture = null;
+                this.dirty = true;
+                this.tileTexture = renderer.createTexture2D(0, gl.LINEAR, gl.LINEAR, gl.REPEAT, gl.REPEAT, gl.RGBA, _this.canvasBuffer, _this.potWidth, _this.potHeight);
+            }, this);
+        }
     },
 
     /**