+++ CHANGELOG.md
@@ -22,6 +22,8 @@
 * The class GameObject has now been exposed, available via `Phaser.GameObjects.GameObject` (thanks @rexrainbow)
 * A Camera following a Game Object will now take the zoom factor of the camera into consideration when scrolling. Fix #3353 (thanks @brandonvdongen)
 * Calling `setText` on a BitmapText object will now recalculate its display origin values. Fix #3350 (thanks @migiyubi)
+* You can now pass an object to Loader.atlas, like you you can with images. Fix #3268 (thanks @TCatshoek)
+* The `onContextRestored` callback won't be defined any more unless the WebGL Renderer is in use in the following objects: BitmapMask, Static Tilemap, TileSprite and Text. This should allow those objects to now work in HEADLESS mode. Fix #3368 (thanks @16patsle)
 
 ## Version 3.2.0 - Kaori - 5th March 2018